requirements.txt
streamlit
PyPDF2
google-generativeai
python-dotenv
PyMuPDF
fitz
langchain-experimental
python-docx
pymupdf
